requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
python-dotenv==1.0.0
pymongo==4.6.0
qdrant-client==1.7.0
openai==1.3.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
orjson==3.9.10
